+++ ris_lib/ris_yaml_resources.h
@@ -12,58 +12,10 @@
 #include <boost/filesystem.hpp>
 
 namespace ris {
-    class yaml_resources {
-        resource_collection collection;
-        std::string root_path;
+    struct yaml_resources : queryable_resources<yaml_resources> {
+        yaml_resources(std::string const& yaml_path) : queryable_resources<yaml_resources>(yaml_path){}
 
-    public:
-        typedef std::unordered_map<std::string, resource> lookup_t;
-
-        yaml_resources(std::string const& yaml_path) {
-            read_from_file(yaml_path);
-            root_path = boost::filesystem::path(yaml_path).parent_path().generic_string();
-        }
-
-        resource_collection const& resources() const {
-            return collection;
-        }
-
-        lookup_t to_lookup() const {
-            lookup_t lookup;
-            for (auto& res : collection.resources) {
-                lookup[res.name] = res;
-            }
-            return lookup;
-        }
-
-        std::string base_path() const {
-            return root_path;
-        }
-
-        std::string header() const {
-            return collection.header;
-        }
-
-        std::string source() const {
-            return collection.source;
-        }
-
-        std::string namespace_() const {
-            return collection.namespace_;
-        }
-
-        std::string class_() const {
-            return collection.class_;
-        }
-
-    private:
-        template< typename TargetType >
-        void optional_get(TargetType& target, std::string const& key, YAML::Node& node) {
-            if (node[key].IsScalar())
-                target = node[key].as<TargetType>();
-        }
-
-        void read_from_file(std::string const& yaml_path) {
+        static void read_from_file(std::string const& yaml_path, resource_collection& collection) {
             try {
                 auto root_node = YAML::LoadFile(yaml_path);
 
@@ -85,6 +37,12 @@ namespace ris {
             catch (std::exception& e) {
                 throw std::runtime_error(std::string("YAML parsing error in ") + yaml_path + ": " + e.what());
             }
+        }
+    private:
+        template< typename TargetType >
+        static void optional_get(TargetType& target, std::string const& key, YAML::Node& node) {
+            if (node[key].IsScalar())
+                target = node[key].as<TargetType>();
         }
     };
 }
